Fix marx2fits with HRC-I
Before this fix, marx2fits fails with
`*** DetectorType HRC-I not supported.`. That's because when initializing the detector, it tries to read the paraemter file for the Verbose setting. In the context of marx2fits, that parameter file is NULL, which means that the parameter is not availble.
HRC-S and ACIS are not effected, since they follow a different code path.